requirements.txt
attrs==25.3.0
certifi==2025.7.9
h11==0.16.0
idna==3.10
outcome==1.3.0.post0
psycopg2-binary==2.9.10
PySocks==1.7.1
python-dotenv==1.1.1
selenium==4.34.2
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
